Refactor headers in panda/board/drivers/*.h
