--- board/drivers/bootkick_declarations.h
@@ -1,5 +1,0 @@
-#pragma once
-
-extern bool bootkick_reset_triggered;
-
-void bootkick_tick(bool ignition, bool recent_heartbeat);
